Update README.md
The project focused on developing a machine learning system that could predict a person's risk of being involved in an accident based on certain features. To achieve this, a neural network was used to analyze data related to the individual's age, gender, and driving history, among other factors. By training the neural network using a suitable architecture and loss function, it was possible to accurately predict the risk of accidents based on these features. The end result was a system that could help identify individuals who may be at a higher risk of accidents, allowing for preventative measures to be taken.